Rename module to component
